pyproject.toml
[project]
name = "langbot"
version = "0.1.0"
description = "LangChain-based personal AI assistant framework"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "LangBot Contributors" }
]
keywords = ["ai", "assistant", "langchain", "agent", "chatbot"]

dependencies = [
    # LangChain 核心
    "langchain>=1.2.13",
    "langchain-core>=1.2.20",
    "langchain-anthropic>=1.4.0",
    "langchain-google-genai>=4.2.1",
    # LangGraph
    "langgraph>=1.1.3",
    "langgraph-checkpoint>=4.0.1",
    "langgraph-prebuilt>=1.0.8",
    "langgraph-sdk>=0.3.12",
    # DeepAgents
    "deepagents>=0.4.11",
    # MCP 适配器
    "langchain-mcp-adapters>=0.2.2",
    "mcp>=1.26.0",
    # LangSmith
    "langsmith>=0.7.22",
    # 配置和 CLI
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "typer>=0.20.0",
    # 日志和 UI
    "loguru>=0.7.3",
    "rich>=14.0.0",
    # HTTP 和 WebSocket
    "httpx>=0.28.1",
    "websockets>=16.0",
    "aiohttp>=3.11.0",
    # 工具依赖
    "tavily-python>=0.5.0",
    "croniter>=6.0.0",
    "readability-lxml>=0.8.4",
    # 通道依赖
    "qq-botpy>=1.2.0",
    # 其他
    "prompt-toolkit>=3.0.52",
    "pyyaml>=6.0.3",
    # 开发依赖
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.1.0",
    "langchain-deepseek>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.3",
]

[project.scripts]
langbot = "langbot.__main__:app"

[project.entry-points."langbot.channels"]
qq = "langbot.channels.qq:QQChannel"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["langbot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
